refactor: dev와 맞춰 redis 패키지 수정
