+++ backend/src/main/java/com/ody/route/service/RouteClientCircuitBreaker.java
@@ -1,6 +1,6 @@
 package com.ody.route.service;
 
-import com.ody.common.redis.CustomRedisTemplate;
+import com.ody.route.repository.RouteClientRedisTemplate;
 import com.ody.route.domain.RouteClientKey;
 import java.time.Duration;
 import java.time.LocalDateTime;
@@ -17,7 +17,7 @@ public class RouteClientCircuitBreaker {
     public static final Duration FAIL_MINUTES_TTL = Duration.ofMinutes(31); // 지연 시간 고려해 31분으로 설정
     public static final Duration BLOCK_HOUR_TTL = Duration.ofHours(3);
 
-    private final CustomRedisTemplate redisTemplate;
+    private final RouteClientRedisTemplate redisTemplate;
 
     public void recordFailCountInMinutes(RouteClient routeClient) {
         String failClientKey = RouteClientKey.getFailKey(routeClient);
